+++ vue/components/ui/atoms/container/container.vue
@@ -77,6 +77,14 @@
     margin: 0px 0px 0px 0px;
 }
 
+.container-ripe.container-ripe-inline {
+    background: none;
+    border: none;
+    box-shadow: none;
+    margin: 0px 0px 0px 0px;
+    transition: none;
+}
+
 body.mobile .container-ripe {
     border-radius: 0px 0px 0px 0px;
     box-shadow: none;
@@ -186,6 +194,7 @@ export const Container = {
             const base = {};
             let maxWidth = this.maxWidth;
             if (this.mode === "expanded") maxWidth = "100%";
+            if (this.mode === "inline") maxWidth = null;
             if (this.isTabletWidth() && this.maxWidthTable) maxWidth = this.maxWidthTablet;
             if (this.isMobileWidth() && this.maxWidthMobile) maxWidth = this.maxWidthMobile;
             if (maxWidth && typeof maxWidth === "number") {